% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_guides.R
\name{render_guides}
\alias{render_guides}
\title{Render all helper modals, introduction, and notes documents}
\usage{
render_guides(.output_dir = "inst/app/www", .actor_id = Sys.getenv("USER"))
}
\arguments{
\item{.output_dir}{character, directory location for output markdowns}

\item{.actor_id}{character, username or identifier. Default is Sys.getenv("USER").}
}
\value{
nothing
}
\description{
Render all helper modals, introduction, and notes documents
}
